# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"eRTMAC-NWIS API","description":"Nearby Wells Intelligence System for Drilling Operations","version":"0.1.0"},"paths":{"/api/health":{"get":{"tags":["health"],"summary":"Check Health","operationId":"check_health_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthCheck"}}}}}}},"/api/wells":{"get":{"tags":["data"],"summary":"Get Wells","operationId":"get_wells_api_wells_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/WellOut"},"type":"array","title":"Response Get Wells Api Wells Get"}}}}}}},"/api/wells/nearby":{"get":{"tags":["data"],"summary":"Get Nearby Wells","operationId":"get_nearby_wells_api_wells_nearby_get","parameters":[{"name":"lat","in":"query","required":true,"schema":{"type":"number","title":"Lat"}},{"name":"lng","in":"query","required":true,"schema":{"type":"number","title":"Lng"}},{"name":"radius_km","in":"query","required":false,"schema":{"type":"number","default":10.0,"title":"Radius Km"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NearbyWellOut"},"title":"Response Get Nearby Wells Api Wells Nearby Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/wells/{well_id}":{"get":{"tags":["data"],"summary":"Get Well","operationId":"get_well_api_wells__well_id__get","parameters":[{"name":"well_id","in":"path","required":true,"schema":{"type":"integer","title":"Well Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WellOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/events":{"get":{"tags":["data"],"summary":"Get Events","operationId":"get_events_api_events_get","parameters":[{"name":"well_id","in":"query","required":false,"schema":{"type":"integer","title":"Well Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DrillingEventOut"},"title":"Response Get Events Api Events Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/events/{event_id}":{"get":{"tags":["data"],"summary":"Get Event","operationId":"get_event_api_events__event_id__get","parameters":[{"name":"event_id","in":"path","required":true,"schema":{"type":"integer","title":"Event Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DrillingEventOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/formations":{"get":{"tags":["data"],"summary":"Get Formations","operationId":"get_formations_api_formations_get","parameters":[{"name":"well_id","in":"query","required":false,"schema":{"type":"integer","title":"Well Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FormationOut"},"title":"Response Get Formations Api Formations Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/documents":{"get":{"tags":["data"],"summary":"Get Documents","operationId":"get_documents_api_documents_get","parameters":[{"name":"well_id","in":"query","required":false,"schema":{"type":"integer","title":"Well Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DocumentOut"},"title":"Response Get Documents Api Documents Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/alerts/lookahead/{well_id}":{"get":{"tags":["intelligence"],"summary":"Get Lookahead Alerts","operationId":"get_lookahead_alerts_api_alerts_lookahead__well_id__get","parameters":[{"name":"well_id","in":"path","required":true,"schema":{"type":"integer","title":"Well Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/depth-correlation/{well_id}":{"get":{"tags":["intelligence"],"summary":"Get Correlation","operationId":"get_correlation_api_depth_correlation__well_id__get","parameters":[{"name":"well_id","in":"path","required":true,"schema":{"type":"integer","title":"Well Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"DocumentOut":{"properties":{"filename":{"type":"string","title":"Filename"},"document_type":{"type":"string","title":"Document Type"},"date":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Date"},"page":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Page"},"extracted_text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Extracted Text"},"id":{"type":"integer","title":"Id"},"well_id":{"type":"integer","title":"Well Id"}},"type":"object","required":["filename","document_type","id","well_id"],"title":"DocumentOut"},"DrillingEventOut":{"properties":{"event_type":{"type":"string","title":"Event Type"},"depth_from":{"type":"number","title":"Depth From"},"depth_to":{"type":"number","title":"Depth To"},"formation_name":{"type":"string","title":"Formation Name"},"severity":{"type":"string","title":"Severity"},"description":{"type":"string","title":"Description"},"cause":{"type":"string","title":"Cause"},"mitigation":{"type":"string","title":"Mitigation"},"outcome":{"type":"string","title":"Outcome"},"id":{"type":"integer","title":"Id"},"well_id":{"type":"integer","title":"Well Id"},"source_document_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Source Document Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["event_type","depth_from","depth_to","formation_name","severity","description","cause","mitigation","outcome","id","well_id","created_at"],"title":"DrillingEventOut"},"FormationOut":{"properties":{"name":{"type":"string","title":"Name"},"depth_from":{"type":"number","title":"Depth From"},"depth_to":{"type":"number","title":"Depth To"},"id":{"type":"integer","title":"Id"},"well_id":{"type":"integer","title":"Well Id"}},"type":"object","required":["name","depth_from","depth_to","id","well_id"],"title":"FormationOut"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthCheck":{"properties":{"status":{"type":"string","title":"Status"},"service":{"type":"string","title":"Service"}},"type":"object","required":["status","service"],"title":"HealthCheck"},"NearbyWellOut":{"properties":{"name":{"type":"string","title":"Name"},"field":{"type":"string","title":"Field"},"latitude":{"type":"number","title":"Latitude"},"longitude":{"type":"number","title":"Longitude"},"spud_date":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Spud Date"},"completion_date":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Completion Date"},"total_depth":{"type":"number","title":"Total Depth"},"current_depth":{"type":"number","title":"Current Depth"},"well_type":{"type":"string","title":"Well Type"},"trajectory":{"type":"string","title":"Trajectory"},"status":{"type":"string","title":"Status"},"is_synthetic":{"type":"boolean","title":"Is Synthetic","default":true},"id":{"type":"integer","title":"Id"},"distance_km":{"type":"number","title":"Distance Km"},"event_count":{"type":"integer","title":"Event Count"}},"type":"object","required":["name","field","latitude","longitude","total_depth","current_depth","well_type","trajectory","status","id","distance_km","event_count"],"title":"NearbyWellOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WellOut":{"properties":{"name":{"type":"string","title":"Name"},"field":{"type":"string","title":"Field"},"latitude":{"type":"number","title":"Latitude"},"longitude":{"type":"number","title":"Longitude"},"spud_date":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Spud Date"},"completion_date":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Completion Date"},"total_depth":{"type":"number","title":"Total Depth"},"current_depth":{"type":"number","title":"Current Depth"},"well_type":{"type":"string","title":"Well Type"},"trajectory":{"type":"string","title":"Trajectory"},"status":{"type":"string","title":"Status"},"is_synthetic":{"type":"boolean","title":"Is Synthetic","default":true},"id":{"type":"integer","title":"Id"}},"type":"object","required":["name","field","latitude","longitude","total_depth","current_depth","well_type","trajectory","status","id"],"title":"WellOut"}}}}